CMR Query and Config Dictionary
Updated configuration files to define a dictionary. This makes it
easy to pass configuration internally instead of relying on global
scope variables.

Added Common Metadata Repository Query option. This allows the user
to specify a daterange and short_name instead of inpaths. The program
will use a web-based CMR query to identify the appropriate inpaths.