Fixing pressure2sigma
Previously incorrectly treated pressure top from OMI. The current implementation
is correct, but is unnecessarily complex. It creates a pressure unit from the model,
then calculates a quasi sigma coordinate for OMI, then converts the pressure value
from the model to OMI-sigma value, then interpolates to that value. This should be
simplified. When OMI is pure pressure, the model pressure should be calculated
and OMI interpolated to that.