Adding OMPROFOZ AvgKern support
This requires allowing the vertical dimension to flip, and it requires diagnosing whether
pressure is a center value or an edge value. OMPROFOZ is an edge value, and OMNO2d is a center
value.

The edge value is diagnosed by the nLevelEdges dimension.
The flip functionality is implemented by flipdims, which is a list of keys to reverse.