Adding scale processor
Some data is stored as shorts or integers, but had Scale and Offset
factors. The netcdf4 library applies scaling only for CF-compliant
scale_factor and add_offset, so these were being missed. This quick
function applies ScaleFactor and Offset only when it is important
(i.e, not 1 and 0, or would change the data type).