cmaqsatproc v2
The first version of cmaqsatproc was useful, but very hard
to update. The new version uses geopandas to bring in more
general functionality that can be applied to any satellite
product that has a pixel polygon or centroid. This allows
for the addition of fractional area overlap calculations.

There is also a move toward a modular design that will
be easier to mix and match. At the core is a satellite
product reader. The reader creates geopandas dataframes
and knows how to calculate weights and weighted contributions
to arbitrary target geopandas.

The goal is to produce gridded data for CMAQ in 2D and 3D.
At this point, the CMAQ-derived AMF has not been calculated.
However, the way this has been designed it should be dramatically
easier to do and to check.

To do:
 - add CMAQ AMF functionality.